add active function to repository

+++ src/Core/Generators/RepositoryGenerator.php
@@ -12,7 +12,7 @@ class RepositoryGenerator implements GeneratorInterface
 {
     public string $baseClass     = 'App\Base\Repositories\BaseEloquentRepository';
     public string $baseInterface = 'App\Base\Interfaces\DataProviderInterface';
-    public array $traits = ['App\Base\Traits\DataProviderTrait'];
+    public array  $traits        = ['App\Base\Traits\DataProviderTrait'];
 
     public function __construct(public GeneratorForm $generatorForm)
     {
@@ -26,25 +26,29 @@ class RepositoryGenerator implements GeneratorInterface
         $path = $this->generatorForm->mainPath . '/Core/Templates/Classes';
         View::addLocation($path);
         View::addNamespace('repository', $path);
-        $renderedModel = View::make('repository')->with(
-            [
-                'generator' => $this,
-            ]);
-        $filename = "{$this->generatorForm->resourceName}Repository.php";
-        $path = base_path(lcfirst($namespace));
-        if (!File::isDirectory($path)) {
-            File::makeDirectory($path, 0777, true, true);
-        }
-
-        if (!File::exists($path . '\\' . $filename) || $this->generatorForm->force) {
-            File::delete($path . '\\' . $filename);
-            if (File::put($path . '\\' . $filename, $renderedModel) !== false) {
-                ConsoleHelper::info('Repository generated! Path in app: ' . lcfirst($namespace) . '\\' . $filename);
-            } else {
-                ConsoleHelper::error('Repository generate error!');
+        try {
+            $renderedModel = View::make('repository')->with(
+                [
+                    'generator' => $this,
+                ]);
+            $filename = "{$this->generatorForm->resourceName}Repository.php";
+            $path = base_path(lcfirst($namespace));
+            if (!File::isDirectory($path)) {
+                File::makeDirectory($path, 0777, true, true);
             }
-        } else {
-            ConsoleHelper::warning('Repository is exists! Add --force option to overwrite Repository!');
+            if (($renderedModel && !File::exists($path . '\\' . $filename)) || $this->generatorForm->force) {
+                File::delete($path . '\\' . $filename);
+                if (File::put($path . '\\' . $filename, $renderedModel) !== false) {
+                    ConsoleHelper::info('Repository generated! Path in app: ' . lcfirst($namespace) . '\\' . $filename);
+                } else {
+                    ConsoleHelper::error('Repository generate error!');
+                }
+            } else {
+                ConsoleHelper::warning('Repository is exists! Add --force option to overwrite Repository!');
+            }
+        } catch (\Throwable $e) {
+            ConsoleHelper::error('Repository generate error!');
+            dd($e);
         }
     }
 }
